refactor all products component
