+++ client/components/index.js
@@ -6,6 +6,6 @@
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {default as ProductCard} from './product-card'
-export {default as AllProductsGrid} from './all-products-grid'
+export {default as ProductsGrid} from './products-grid'
 export {Login, Signup} from './auth-form'
 export {default as FilterForm} from './filter-form'
